pyproject.toml
[project]
name = "python-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=46.0.3",
    "dcap-qvl>=0.3.8",
    "fastapi>=0.128.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
